refactor(ui): move chat page helpers to owning modules
