+++ ui/src/pages/chat/split-layout.ts
@@ -12,6 +12,16 @@ export type ChatSplitLayout = {
 
 const MIN_PAIR_SHARE = 0.15;
 
+export function splitWeight(weights: number[], index: number, context: string): number {
+  return expectDefined(weights[index], context);
+}
+
+export function splitRatio(weights: number[], index: number, context: string): number {
+  const before = splitWeight(weights, index, `${context} before divider`);
+  const after = splitWeight(weights, index + 1, `${context} after divider`);
+  return before / (before + after);
+}
+
 function cloneLayout(layout: ChatSplitLayout): ChatSplitLayout {
   return {
     columns: layout.columns.map((column) => ({
